Editor: Fixed crash caused by phantom element with dead pointer Because was registered element with ITEM_LAST_SIZE equal to zero, then is no way to remove it since a logical condition prevents unregistering of zero-size entries. Also, for better accuracy, use 64-bit integer based coordinate unit since all elements are aligning with 1-pixel grid even grid snapping is disabled

+++ Editor/editing/_scenes/level/lvl_scene.h
@@ -352,8 +352,8 @@ public:
     QGraphicsItem *itemCollidesWith(QGraphicsItem *item, PGE_ItemList *itemgrp = 0);
     QGraphicsItem *itemCollidesCursor(QGraphicsItem *item);
 
-    typedef RTree<QGraphicsItem *, double, 2, double > IndexTree;
-    typedef double RPoint[2];
+    typedef RTree<QGraphicsItem *, int64_t, 2, int64_t > IndexTree;
+    typedef int64_t RPoint[2];
     IndexTree tree;
     void queryItems(QRectF &zone, PGE_ItemList *resultList);
     void queryItems(double x, double y, PGE_ItemList *resultList);
